requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
psycopg2-binary==2.9.9
yfinance>=1.3.0
curl_cffi
matplotlib==3.9.2
beautifulsoup4==4.12.3
weasyprint>=66
pydyf>=0.11
